# v1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Micelios API",
        "description": "API de comunicación micelios backend",
        "version": "1.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/categoria/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Mostrar la información de una categoria pasando un ID",
                "operationId": "f4d993e57f9317d7ab8850ea2e6c07de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Categoria ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "category_name": {
                                            "type": "string",
                                            "example": "nombre de categoria"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se ha encontrado información disponible"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/producto/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar la información de un producto pasando un ID",
                "operationId": "92b9254a62bbb42b6221260e2a7f9841",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "categoria_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "nombre_producto": {
                                            "type": "string",
                                            "example": "Nombre Producto"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}